requirements.txt
openai
python-dotenv
python-docx
langchain
langchain-openai
langchain-community
chromadb
streamlit
watchdog
# pydantic
